Project now uses static inventory rather than dynamic.
This is needed as the EC2 dynamic inventory wasn't able to keep track of additional variables and metadata that we need to record, e.g. `ansible_python_interpreter`. As part of this, some helper variables were removed or restructured such that the inventory is used, instead.

Incidentally, I also fixed some bugs with the Jenkins configuration that were making it hard to test all of these other changes.